requirement.txt
requests==2.27.1
scikit-learn==1.0.2
timm==0.6.1
torch==1.10.2
transformers==4.14.1
torchvision==0.11.3
tqdm==4.63.0
tensorboard==2.10.1
tensorboard-logger==0.1.0
tnesorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
scikit-learn==1.0.2
scipy==1.7.3
pandas==1.4.1
pillow==9.0.1